requirements.txt
anyio==4.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
notion-client==2.2.1
numpy==1.26.3
oauthlib==3.2.2
pandas==2.1.4
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
watchdog==4.0.0
